fix: main file and file names
